refactor: force output to consistent naming scheme
